Rework duplicates handling
We won't mangle names of the attributes, on a duplicate, we will mangle the
names of the elements instead. Idea behind this change is that with
xsd:extension you may extend the same type multiple times and hence keep the
same attributes around. In that case users of go code are better served with
attributes always keeping the attribute names consistent.

Obviously, you could come up with the vice versa scenario where the opposite is
beneficial, however I made this design choice at this point anyway.