Introduce Choice.Elements() method
This mimics the interface of Sequence. Likewise in Sequence we will have
Elements() method to return all the elements and ElementList object variable to
note the direct Elements.

This change is not necessary at this point, but achieving consistency with
sequence may turn out beneficial in future as the complexity of the project
raises.