Bump go version
Let's bump our go version requirement a bit to allow dependencies to be
updated a bit. This is perhaps overly conservative bump since the 1.15
release is only supported by some long life linuxes and not the golang
upstream, but there is no rush as 1.16 and 1.17 does not bring anything
particular we need at the moment.

I will be happily challenged on this.